requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.11.12
cffi==2.0.0
dnspython==2.8.0
minio==7.2.20
neo4j==6.0.3
psycopg2==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pymongo==4.15.5
python-dotenv==1.2.1
pytz==2025.2
typing_extensions==4.15.0
urllib3==2.6.0
